refactor: 자동차관리 라운드당 레이스 메서드 수정
